% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_clustering_quality.R
\name{evaluate_clustering_quality}
\alias{evaluate_clustering_quality}
\title{Evaluate Clustering Quality}
\usage{
evaluate_clustering_quality(D, group)
}
\arguments{
\item{D}{A dissimilarity matrix or a `dist` object.}

\item{group}{A vector of true class labels for samples (used for label alignment).}
}
\value{
A list with the following elements:
\describe{
  \item{avg_silhouette}{Average silhouette width. Higher is better (ideal: > 0.7).}
  \item{dbi}{Davies-Bouldin Index. Lower is better (ideal: < 0.5).}
  \item{chi}{Calinski-Harabasz Index. Higher is better (ideal: > 500).}
  \item{dunn}{Dunn Index. Higher is better (ideal: > 1.5).}
}
}
\description{
This function evaluates the quality of a clustering result based on a dissimilarity matrix
using several internal validation metrics: silhouette score, Davies-Bouldin Index,
Calinski-Harabasz Index, and Dunn Index. It uses k-medoids clustering (`pam`) to perform clustering.
}
