\name{snijders}
\docType{data}
\alias{gm00143}
\alias{gm01524}
\alias{gm01535}
\alias{gm01750}
\alias{gm02948}
\alias{gm03134}
\alias{gm03563}
\alias{gm03576}
\alias{gm04435}
\alias{gm05296}
\alias{gm07081}
\alias{gm07408}
\alias{gm10315}
\alias{gm13031}
\alias{gm13330}
\title{Public CGH data of Snijders}
\description{
The data consist of 15 human cell strains with known karyotype (12 fibroblast cell strains, 2 chorionic villus cell strains, 1 lymploblast cell strain) from the NIGMS Human Genetics Cell Repository (http://locus.umdnj.edu/nigms). Each cell strain has been hybridized onto a CGH-array of 2276 BAC's spotted in triplicate.
}
\usage{data(snijders)}
\references{
  A M Snijders, N Nowak, R Segraves, S Blackwood, N Brown, J
    Conroy, G Hamilton, A K Hindle, B Huey, K Kimura, S Law, K Myambo, J
    Palmer, B Ylstra, J P Yue, J W Gray, A N Jain, D Pinkel & D G
    Albertson , Assembly of microarrays for genome-wide measurement of
    DNA copy number,
    \emph{Nature Genetics} 29, pp 263 - 264 (2001) Brief Communications.
 }
\source{\url{http://www.nature.com/ng/journal/v29/n3/suppinfo/ng754_S1.html}}
\examples{
data(snijders)
array <- gm13330
}
\keyword{datasets}
