\name{veltman}
\docType{data}
\alias{P9}
\alias{P20}
\title{Public CGH data of Veltman}
\description{
The data consist of 2 bladder cancer tumors obtained by Veltman et al (2003).
}
\usage{data(veltman)}
\references{
  \item{ }{Joris A. Veltman, Jane Fridlyand, Sunanda Pejavar, Adam
    B. Olshen, James E. Korkola, Sandy DeVries, Peter Carroll, Wen-Lin
    Kuo, Daniel Pinkel, Donna Albertson, Carlos Cordon-Cardo, Ajay
    N. Jain and Frederic M. Waldman.
    Array-based Comparative Genomic Hybridization for Genome-Wide
    Screening of DNA Copy Number in Bladder Tumors.
    Cancer Research 63, 2872-2880, 2003.
  }
 }
\source{\url{http://cancerres.aacrjournals.org/cgi/content/full/63/11/2872}}
\examples{
data(veltman)
P9
}
\keyword{datasets}

