\name{GLAD-internal}
\title{GLAD-internal}
\alias{GLAD-internal}
\alias{affectationGNL}
\alias{affectationGNL.profileCGH}
\alias{chrBreakpoints}
\alias{chrBreakpoints.profileCGH}
\alias{cluster}
\alias{cluster.default}
\alias{detectOutliers}
\alias{detectOutliers.profileChr}  
\alias{filterBkp}
\alias{filterBkp.profileCGH}
\alias{findCluster}
\alias{findCluster.profileChr}
\alias{lawsglad}
\alias{loopRemove}
\alias{loopRemove.profileChr}
\alias{MoveBkp}
\alias{MoveBkp.profileCGH}
\alias{OutliersGNL}
\alias{OutliersGNL.profileCGH}
\alias{plotCytoBand}
\alias{plotCytoBand.default}
\alias{removeBreakpoints}
\alias{removeBreakpoints.profileChr}
\alias{removeLevel}
\alias{removeLevel.profileChr}
\alias{MoveBkpStep}
\alias{MoveBkpStep.profileCGH}
\alias{testBkpToMove}
\alias{testBkpToMove.profileCGH}
\alias{filterBkpStep}
\alias{filterBkpStep.profileCGH}
\alias{BkpInfo}
\alias{BkpInfo.profileCGH}


\usage{ }

\description{ }

\value{ }

\author{Philippe Hupé, \url{glad@curie.fr}.}

\seealso{}

\keyword{internal}



